# docs/swagger.yaml
definitions:
  domain.Song:
    properties:
      group:
        description: Название группы
        type: string
      id:
        description: Уникальный идентификатор песни
        type: integer
      link:
        description: Ссылка на дополнительную информацию
        type: string
      release_date:
        description: Дата релиза песни
        type: string
      song:
        description: Название песни
        type: string
      text:
        description: Текст песни
        type: string
    type: object
  domain.SongCreateRequest:
    properties:
      group:
        description: Название группы (обязательно)
        type: string
      song:
        description: Название песни (обязательно)
        type: string
    type: object
  domain.SongDetail:
    properties:
      link:
        description: Ссылка на дополнительную информацию о песне
        type: string
      releaseDate:
        description: Дата релиза песни
        type: string
      text:
        description: Текст песни
        type: string
    type: object
info:
  contact: {}
  description: API для управления библиотекой песен
  termsOfService: http://swagger.io/terms/
  title: Song Library API
  version: "1.0"
paths:
  /info:
    get:
      description: Получение деталей о песне по имени группы и названию песни.
      parameters:
      - description: Название группы
        in: query
        name: group
        required: true
        type: string
      - description: Название песни
        in: query
        name: song
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SongDetail'
        "400":
          description: Параметры обязательны
          schema:
            type: string
        "500":
          description: Ошибка получения данных
          schema:
            type: string
      summary: Получить информацию о песне
      tags:
      - Info
  /library:
    get:
      description: Получение списка песен с фильтрацией по группе, названию и дате
        релиза.
      parameters:
      - default: 1
        description: Номер страницы
        in: query
        name: page
        type: integer
      - default: 10
        description: Количество элементов на странице
        in: query
        name: limit
        type: integer
      - description: Фильтр по группе
        in: query
        name: group
        type: string
      - description: Фильтр по названию песни
        in: query
        name: song
        type: string
      - description: Фильтр по дате релиза
        in: query
        name: release_date
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Song'
            type: array
        "500":
          description: Ошибка получения библиотеки
          schema:
            type: string
      summary: Получить библиотеку песен
      tags:
      - Songs
  /song:
    post:
      description: Добавление новой песни в библиотеку.
      parameters:
      - description: Данные для создания песни
        in: body
        name: song
        required: true
        schema:
          $ref: '#/definitions/domain.SongCreateRequest'
      responses:
        "201":
          description: Песня добавлена
        "400":
          description: Ошибка декодирования данных песни
          schema:
            type: string
        "500":
          description: Ошибка добавления песни
          schema:
            type: string
      summary: Добавить песню
      tags:
      - Songs
  /song/{id}:
    delete:
      description: Удаление песни из библиотеки по ID.
      parameters:
      - description: ID песни
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Песня удалена
        "400":
          description: Неверный ID песни
          schema:
            type: string
        "500":
          description: Ошибка удаления песни
          schema:
            type: string
      summary: Удалить песню
      tags:
      - Songs
    put:
      description: Обновление информации о песне по ID.
      parameters:
      - description: ID песни
        in: path
        name: id
        required: true
        type: integer
      - description: Данные песни
        in: body
        name: song
        required: true
        schema:
          $ref: '#/definitions/domain.Song'
      responses:
        "200":
          description: Песня обновлена
        "400":
          description: Ошибка декодирования данных или неверный ID
          schema:
            type: string
        "500":
          description: Ошибка обновления песни
          schema:
            type: string
      summary: Обновить данные песни
      tags:
      - Songs
  /song/{id}/text:
    get:
      description: Получение текста песни по ID с возможностью пагинации по куплетам.
      parameters:
      - description: ID песни
        in: path
        name: id
        required: true
        type: integer
      - default: 1
        description: Номер страницы (по куплетам)
        in: query
        name: page
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверный ID песни
          schema:
            type: string
        "404":
          description: Песня или куплеты не найдены
          schema:
            type: string
      summary: Получить текст песни
      tags:
      - Songs
swagger: "2.0"